+++ README.md
@@ -12,7 +12,7 @@ he can scan the QR code and send a message directly to the car owner.
 
 
 ## Technologies Used
-- Python
+- Python 3.10
 - Kivy
 - Firebase
 
@@ -21,12 +21,14 @@ he can scan the QR code and send a message directly to the car owner.
 - Sebastian Soczawa
 
 ## dev setup
-
 - git clone https://github.com/Green0x0y/Kokos_App.git 
-- cd myapp
-- virtualenv -p python3 venv
-- source venv/bin/activate
+- cd Kokos_App/myapp
+- python3.10 -m venv kokos-venv
+- source kokos-venv/bin/activate
 - pip install -r requirements.txt
-- python3 ./main.py
+- python3 ./app.py
+
+In order to run the application python 3.10 is required.
+Installing mulitiple python versions can be done for example with **pyenv**
