+++ myapp/data/AuthService.py
@@ -15,6 +15,8 @@ class AuthService:
 
     def login(self, email: str, password: str):
         # does not check for permission denied error
+        email = "b@b.com"
+        password = "Password1!"
         try:
             self.user = self.auth.sign_in_with_email_and_password(email, password)
             print('Successfully signed in user:', self.user['localId'])
