--- myapp/data/requirements.txt
@@ -1,1 +1,0 @@
-pip install pyrebase
