+++ myapp/screens/usercode.py
@@ -1,0 +1,12 @@
+from kivy.uix.screenmanager import Screen
+
+class UserCodeScreen(Screen):
+    def verify_code(self, instance):
+        code = self.ids.code_input.text
+        if len(code) != 8:
+            print("Code must be 8 characters long")
+        elif not code.isalnum():
+            print("Code must contain only letters and numbers")
+        else:
+            # TODO: Handle valid code
+            print("Valid code entered")
